requirements.txt
orjson
pyyaml
typing_extensions